Please write me a message if you use this plugin
If you use this plugin, please write me a short message with a link to the project where you embed the plugin, and what would you like to have more in the plugin. It will help me to stay focused on the important issues and see the global picture! 
It is not mandatory, but i will really appreciate it!